PVAYLADEV-749: Rename artifact, groupId and source code package
