+++ src/src/main/java/fi/vrk/xroad/restadapterservice/ConsumerGateway.java
@@ -20,12 +20,12 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package com.pkrete.restgateway;
+package fi.vrk.xroad.restadapterservice;
 
-import com.pkrete.restgateway.endpoint.ConsumerEndpoint;
-import com.pkrete.restgateway.util.Constants;
-import com.pkrete.restgateway.util.ConsumerGatewayUtil;
-import com.pkrete.restgateway.util.RESTGatewayUtil;
+import fi.vrk.xroad.restadapterservice.endpoint.ConsumerEndpoint;
+import fi.vrk.xroad.restadapterservice.util.Constants;
+import fi.vrk.xroad.restadapterservice.util.ConsumerGatewayUtil;
+import fi.vrk.xroad.restadapterservice.util.RESTGatewayUtil;
 import com.pkrete.xrd4j.client.SOAPClient;
 import com.pkrete.xrd4j.client.SOAPClientImpl;
 import com.pkrete.xrd4j.client.deserializer.AbstractResponseDeserializer;
