+++ src/src/main/java/fi/vrk/xroad/restadapterservice/endpoint/AbstractEndpoint.java
@@ -20,7 +20,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package com.pkrete.restgateway.endpoint;
+package fi.vrk.xroad.restadapterservice.endpoint;
 
 /**
  * This abstract class is a base class for ProviderEndpoint and ConsumerEndpoint
