+++ src/src/main/java/fi/vrk/xroad/restadapterservice/endpoint/ConsumerEndpoint.java
@@ -20,7 +20,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package com.pkrete.restgateway.endpoint;
+package fi.vrk.xroad.restadapterservice.endpoint;
 
 import com.pkrete.xrd4j.common.member.ConsumerMember;
 import com.pkrete.xrd4j.common.member.ProducerMember;
