+++ src/src/test/java/fi/vrk/xroad/restadapterservice/util/ConsumerGatewayUtilTest.java
@@ -20,13 +20,13 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package com.pkrete.restgateway.util;
+package fi.vrk.xroad.restadapterservice.util;
 
 import com.pkrete.xrd4j.common.exception.XRd4JException;
 import com.pkrete.xrd4j.common.member.ConsumerMember;
 import com.pkrete.xrd4j.common.member.ProducerMember;
 import com.pkrete.xrd4j.common.util.SOAPHelper;
-import com.pkrete.restgateway.endpoint.ConsumerEndpoint;
+import fi.vrk.xroad.restadapterservice.endpoint.ConsumerEndpoint;
 import java.io.UnsupportedEncodingException;
 import java.util.Map;
 import java.util.Properties;
