+++ src/src/test/resources/test-log4j.xml
@@ -7,13 +7,13 @@
         </layout>
     </appender>
     <appender name="file" class="org.apache.log4j.FileAppender">
-        <param name="File" value="test-rest-gateway.log"/>
+        <param name="File" value="test-rest-adapter-service.log"/>
         <param name="Append" value="false"/>
         <layout class="org.apache.log4j.PatternLayout">
             <param name="ConversionPattern" value="%d{dd.MM.yyyy HH:mm:ss} %-5p %c{1} : %m%n"/>
         </layout>
     </appender>
-    <!--<logger name="com.pkrete.restgateway" additivity="false">
+    <!--<logger name="fi.vrk.xroad.restadapterservice" additivity="false">
         <level value="DEBUG"/>
         <appender-ref ref="console"/>
     </logger>-->
